Feature/setup ci

+++ src/application/workflow.rs
@@ -1,83 +1,72 @@
-use sqlx::{Pool, Sqlite};
 use anyhow::Result;
+use sqlx::{Pool, Sqlite};
 
 use crate::domain::model::{
-    TodoId, CreateTodoInput, UpdateTodoInput, TodoOutput,
-    create_todo, update_todo, delete_todo, find_todo, list_todos
+    create_todo, delete_todo, find_todo, list_todos, update_todo, CreateTodoInput, TodoId,
+    TodoOutput, UpdateTodoInput,
 };
 use crate::infrastructure::data_access::{
-    find_todo_by_id, find_all_todos, save_todo, delete_todo_by_id, DataAccessError
+    delete_todo_by_id, find_all_todos, find_todo_by_id, save_todo, DataAccessError,
 };
 
 pub async fn create_todo_workflow(
-    pool: &Pool<Sqlite>, 
-    input: CreateTodoInput
+    pool: &Pool<Sqlite>,
+    input: CreateTodoInput,
 ) -> Result<TodoOutput, DataAccessError> {
     let output = create_todo(input);
-    
-    match &output {
-        TodoOutput::Created(todo) => {
-            save_todo(pool, todo).await?;
-        },
-        _ => {},
+
+    if let TodoOutput::Created(todo) = &output {
+        save_todo(pool, todo).await?;
     }
-    
+
     Ok(output)
 }
 
 pub async fn update_todo_workflow(
-    pool: &Pool<Sqlite>, 
-    input: UpdateTodoInput
+    pool: &Pool<Sqlite>,
+    input: UpdateTodoInput,
 ) -> Result<TodoOutput, DataAccessError> {
     let existing_todo = find_todo_by_id(pool, &input.id).await?;
-    
+
     let output = update_todo(input, existing_todo);
-    
-    match &output {
-        TodoOutput::Updated(todo) => {
-            save_todo(pool, todo).await?;
-        },
-        _ => {},
+
+    if let TodoOutput::Updated(todo) = &output {
+        save_todo(pool, todo).await?;
     }
-    
+
     Ok(output)
 }
 
 pub async fn delete_todo_workflow(
-    pool: &Pool<Sqlite>, 
-    id: TodoId
+    pool: &Pool<Sqlite>,
+    id: TodoId,
 ) -> Result<TodoOutput, DataAccessError> {
     let existing_todo = find_todo_by_id(pool, &id).await?;
-    
+
     let output = delete_todo(id.clone(), existing_todo);
-    
-    match &output {
-        TodoOutput::Deleted(id) => {
-            delete_todo_by_id(pool, id).await?;
-        },
-        _ => {},
+
+    if let TodoOutput::Deleted(id) = &output {
+        delete_todo_by_id(pool, id).await?;
     }
-    
+
     Ok(output)
 }
 
 pub async fn find_todo_workflow(
-    pool: &Pool<Sqlite>, 
-    id: TodoId
+    pool: &Pool<Sqlite>,
+    id: TodoId,
 ) -> Result<TodoOutput, DataAccessError> {
     let existing_todo = find_todo_by_id(pool, &id).await?;
-    
+
     let output = find_todo(id, existing_todo);
-    
+
     Ok(output)
 }
 
-pub async fn list_todos_workflow(
-    pool: &Pool<Sqlite>
-) -> Result<TodoOutput, DataAccessError> {
+pub async fn list_todos_workflow(pool: &Pool<Sqlite>) -> Result<TodoOutput, DataAccessError> {
     let todos = find_all_todos(pool).await?;
-    
+
     let output = list_todos(todos);
-    
+
     Ok(output)
 }
